Separate network code from app_main
