+++ main/CMakeLists.txt
@@ -1,4 +1,5 @@
-set(SOURCES main.c)
+file(GLOB_RECURSE SOURCES *.c)
+
 idf_component_register(SRCS ${SOURCES}
                     INCLUDE_DIRS .
                     EMBED_FILES server/error.html server/favicon.ico server/test.css server/test.js server/root.html)
