+++ main/main.c
@@ -39,12 +39,11 @@
 #include "tp_spi.h"
 #include "touch_driver.h"
 
+#include "network_test.h"
 
 /*********************
  *      DEFINES
  *********************/
-
-#define MODE_AP_STA 1  //if this is 0 ESP32 will boot in AP mode, 1 will boot in STA mode
 
 // Detect the use of a shared SPI Bus and verify the user specified the same SPI bus for both touch and tft
 #if (CONFIG_LVGL_TOUCH_CONTROLLER == 1 || CONFIG_LVGL_TOUCH_CONTROLLER == 3) && TP_SPI_MOSI == DISP_SPI_MOSI && TP_SPI_CLK == DISP_SPI_CLK
@@ -66,28 +65,6 @@
 
 #define SHARED_SPI_BUS
 #endif
-
- #define EXAMPLE_ESP_MAXIMUM_RETRY  5
-// FreeRTOS event group to signal when we are connected
-static EventGroupHandle_t s_wifi_event_group;
-/* The event group allows multiple bits for each event, but we only care about two events:
- * - we are connected to the AP with an IP
- * - we failed to connect after the maximum amount of retries */
-#define WIFI_CONNECTED_BIT BIT0
-#define WIFI_FAIL_BIT      BIT1
-
-#if MODE_AP_STA == 0
-#define EXAMPLE_ESP_WIFI_SSID      CONFIG_ESP_AP_WIFI_SSID
-#define EXAMPLE_ESP_WIFI_PASS      CONFIG_ESP_AP_WIFI_PASSWORD
-#define EXAMPLE_MAX_STA_CONN       CONFIG_ESP_AP_MAX_STA_CONN
-static const char *TAG = "wifi AP mode";
-#else
-#define SSID  	CONFIG_ESP_WIFI_SSID
-#define PASS  	CONFIG_ESP_WIFI_PASSWORD
-static const char *TAG = "wifi STA mode";
-#endif
-
-static int s_retry_num = 0;
 
 /**********************
  *  STATIC PROTOTYPES
@@ -116,11 +93,6 @@ static lv_obj_t * cont;
  * @return LV_RES_OK because the object is not deleted in this function
  */
 static void btn_event_cb(lv_obj_t * btn, lv_event_t event);
-
-static void event_handler(void* arg, esp_event_base_t event_base,
-                        int32_t event_id, void* event_data);
-
-void wifi_init_sta(void);
 
 void lv_tutorial_objects(void);
 
@@ -224,110 +196,6 @@ static void configure_shared_spi_bus(void)
   tp_spi_add_device(TOUCH_SPI_HOST);
 }
 #endif
-
-static void event_handler(void* arg, esp_event_base_t event_base,
-                                int32_t event_id, void* event_data)
-{
-    if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_START) {
-        esp_wifi_connect();
-    } else if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_DISCONNECTED) {
-        if (s_retry_num < EXAMPLE_ESP_MAXIMUM_RETRY) {
-            esp_wifi_connect();
-            s_retry_num++;
-            ESP_LOGI(TAG, "retry to connect to the AP");
-        } else {
-            xEventGroupSetBits(s_wifi_event_group, WIFI_FAIL_BIT);
-        }
-        ESP_LOGI(TAG,"connect to the AP fail");
-    } else if (event_base == IP_EVENT && event_id == IP_EVENT_STA_GOT_IP) {
-        ip_event_got_ip_t* event = (ip_event_got_ip_t*) event_data;
-        ESP_LOGI(TAG, "got ip:%s", ip4addr_ntoa(&event->ip_info.ip));
-        s_retry_num = 0;
-        xEventGroupSetBits(s_wifi_event_group, WIFI_CONNECTED_BIT);
-    }
-    if (event_id == WIFI_EVENT_AP_STACONNECTED) {
-        wifi_event_ap_staconnected_t* event = (wifi_event_ap_staconnected_t*) event_data;
-        ESP_LOGI(TAG, "station "MACSTR" join, AID=%d",
-                 MAC2STR(event->mac), event->aid);
-    } else if (event_id == WIFI_EVENT_AP_STADISCONNECTED) {
-        wifi_event_ap_stadisconnected_t* event = (wifi_event_ap_stadisconnected_t*) event_data;
-        ESP_LOGI(TAG, "station "MACSTR" leave, AID=%d",
-                 MAC2STR(event->mac), event->aid);
-    }
-}
-
-void wifi_init_sta(void)
-{
-    s_wifi_event_group = xEventGroupCreate();
-
-    tcpip_adapter_init();
-
-    ESP_ERROR_CHECK(esp_event_loop_create_default());
-
-    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
-    ESP_ERROR_CHECK(esp_wifi_init(&cfg));
-
-    ESP_ERROR_CHECK(esp_event_handler_register(WIFI_EVENT, ESP_EVENT_ANY_ID, &event_handler, NULL));
-    ESP_ERROR_CHECK(esp_event_handler_register(IP_EVENT, IP_EVENT_STA_GOT_IP, &event_handler, NULL));
-
-#if MODE_AP_STA == 0
-	wifi_config_t ap_wifi_config = {
-        .ap = {
-            .ssid = EXAMPLE_ESP_WIFI_SSID,
-            .ssid_len = strlen(EXAMPLE_ESP_WIFI_SSID),
-            .password = EXAMPLE_ESP_WIFI_PASS,
-            .max_connection = EXAMPLE_MAX_STA_CONN,
-            .authmode = WIFI_AUTH_WPA_WPA2_PSK
-        },
-    };
-
-    if (strlen(EXAMPLE_ESP_WIFI_PASS) == 0) {
-        ap_wifi_config.ap.authmode = WIFI_AUTH_OPEN;
-    }
-    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_AP) ); 
-    ESP_ERROR_CHECK(esp_wifi_set_config(ESP_IF_WIFI_AP, &ap_wifi_config) );
-#else
-	wifi_config_t sta_wifi_config = {
-        .sta = {
-            .ssid = SSID,
-            .password = PASS
-        },
-    };
-    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_STA) ); 
-    ESP_ERROR_CHECK(esp_wifi_set_config(ESP_IF_WIFI_STA, &sta_wifi_config) );
-#endif
-    
-    ESP_ERROR_CHECK(esp_wifi_start() );
-
-#if MODE_AP_STA == 0
-    ESP_LOGI(TAG, "Soft AP initialised. SSID:%s password:%s",
-        EXAMPLE_ESP_WIFI_SSID, EXAMPLE_ESP_WIFI_PASS);
-#else
-	/* Waiting until either the connection is established (WIFI_CONNECTED_BIT) or connection failed for the maximum
-     * number of re-tries (WIFI_FAIL_BIT). The bits are set by event_handler() (see above) */
-    EventBits_t bits = xEventGroupWaitBits(s_wifi_event_group,
-            WIFI_CONNECTED_BIT | WIFI_FAIL_BIT,
-            pdFALSE,
-            pdFALSE,
-            portMAX_DELAY);
-
-    /* xEventGroupWaitBits() returns the bits before the call returned, hence we can test which event actually
-     * happened. */
-    if (bits & WIFI_CONNECTED_BIT) {
-        ESP_LOGI(TAG, "connected to ap SSID:%s password:%s",
-                 SSID, PASS);
-    } else if (bits & WIFI_FAIL_BIT) {
-        ESP_LOGI(TAG, "Failed to connect to SSID:%s, password:%s",
-                 SSID, PASS);
-    } else {
-        ESP_LOGE(TAG, "UNEXPECTED EVENT");
-    }
-    vEventGroupDelete(s_wifi_event_group);
-#endif
-    
-    ESP_ERROR_CHECK(esp_event_handler_unregister(IP_EVENT, IP_EVENT_STA_GOT_IP, &event_handler));
-    ESP_ERROR_CHECK(esp_event_handler_unregister(WIFI_EVENT, ESP_EVENT_ANY_ID, &event_handler));   
-}
 
 void lv_tutorial_objects(void)
 {
